sonar test

+++ sw/airborne/subsystems/ins/ins_int.h
@@ -61,7 +61,7 @@ struct InsInt {
 
 #if USE_SONAR
   bool_t update_on_agl; ///< use sonar to update agl if available
-  int sonar_z;
+  float sonar_z;
 #endif
 };
